Add Table & Time components & Objects file
